clean up even more code
