+++ nextjs/src/app/api/profile/route.ts
@@ -1,7 +1,7 @@
 "use server";
 import { createServiceRoleClient } from "@/utils/supabase/server";
 import { NextResponse } from "next/server";
-import initClients from "@/lib/initClient";
+import { initClients } from "@/utils/initClients";
 
 async function getProfile(supabase: ReturnType<typeof createServiceRoleClient>, userId: string) {
     const { data, error } = await supabase.from("profiles").select("*").eq("id", userId);
